Add a new test for timezone conversions
Add two more dates, each 1 second before the existing date, so it is in
the previous month. Two of the dates are in DST, and the other two are
in ST. If there are any shifts because of DST conversions, then this
should detect it.

(cherry picked from commit 79ff88c1e08d7bb770215e41f7df053a39f0131a)